1846 Allow creation of the comment root without a request
In order to initialize the root comment for a particular object, there
needs to be a way to call _get_or_create_tree_root without a request.

Without this, you would not be able to create comment objects
programmatically without first having a request ping the load_comments
url with the correct GET params.